requirements.txt
# cogitto-medication-ai/requirements.txt
# Cogitto: Medication AI Assistant Dependencies

# Core Framework (Current)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Development
pytest==7.4.3
httpx==0.25.2

# Future Dependencies (Phase 2+)
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
# redis==5.0.1
# openai==1.3.8
# pandas==2.1.3
# aiohttp==3.10.0



# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Environment management
python-dotenv==1.0.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
